🧪 Test teleportation sampling in C++
Reuse native benchmark test helpers and keep the Python test focused on bindings.
